Refactor to include AbstractRetryApiClient
